requirements.txt
streamlit
google_generativeai
python-dotenv
langchain
PyPDF2
langchain_community
langchain_google_genai
-e .